feat(YB-84): 레포지토리 구성 및 수정

+++ yellobook-api/src/main/java/com/yellobook/domain/inventory/service/InventoryQueryService.java
@@ -4,7 +4,7 @@ import com.yellobook.domain.inventory.dto.GetProductsResponse;
 import com.yellobook.domain.inventory.dto.GetTotalInventoryResponse;
 
 public interface InventoryQueryService {
-    public GetTotalInventoryResponse getTotalInventory(Long page, Long size, Long teamId);
+    public GetTotalInventoryResponse getTotalInventory(Integer page, Integer size, Long teamId);
     public GetProductsResponse getProductsByInventory(Long teamId, Long inventoryId);
 
     public GetProductsResponse getProductByKeywordAndInventory(Long teamId, Long inventoryId, String keyword);
